Rename xor_name module to xor_space
