+++ src/action.rs
@@ -6,10 +6,7 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use crate::error::InterfaceError;
-use crate::routing_table::Authority;
-use crate::xor_name::XorName;
-use crate::NetworkBytes;
+use crate::{error::InterfaceError, routing_table::Authority, xor_space::XorName, NetworkBytes};
 use hex_fmt::HexFmt;
 use quic_p2p::Token;
 use std::fmt::{self, Debug, Formatter};
