+++ src/utils.rs
@@ -6,7 +6,7 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use crate::xor_name::XorName;
+use crate::xor_space::XorName;
 use std::{
     fmt::{self, Display, Formatter},
     ops::RangeInclusive,
