+++ src/xor_space/xorable.rs
